+++ R/pretty_inline_ci.R
@@ -3,6 +3,8 @@
 #' and adds a user specified explanatory note; to be used inline.
 #' @param string A character vector as formated by \code{\link[prettypublisher]{pretty_ci}}.
 #' @param note A character vector indicating the explanatory note to be used.
+#' @param replace_bracket Logical, defaults to \code{TRUE}. Should the leading bracket be automatically
+#' be reinserted in front of the custom note supplied by \code{note}.
 #' @importFrom stringr str_replace
 #' @seealso pretty_ci
 #' @return An estimate with formated CI's for use inline within a document.
@@ -23,9 +25,13 @@
 #' formated_ci <- pretty_ci(est, lci, uci, sep = '-')
 #' pretty_inline_ci(formated_ci)
 #'
-pretty_inline_ci <- function(string, note = "95% CI ") {
+#'## Add a note before the braket
+#'pretty_inline_ci(formated_ci, note = " per 100,000 (95% CI ", replace_bracket = FALSE)
+pretty_inline_ci <- function(string, note = "95% CI ", replace_bracket = TRUE) {
 
-  note <- paste0("(", note)
+  if (replace_bracket) {
+    note <- paste0("(", note)
+  }
 
   string <- string %>%
               str_replace("\\(", note)